client: first stab at peering refeactor [no ci]
