+++ packages/client/src/net/peer/peer.ts
@@ -1,8 +1,8 @@
+import { BIGINT_NEG1 } from '@ethereumjs/util'
 import { EventEmitter } from 'events'
 
 import type { Config } from '../../config'
 import type {
-  BoundProtocol,
   EthProtocolMethods,
   LesProtocolMethods,
   Protocol,
@@ -38,13 +38,12 @@ export interface PeerOptions {
  * Network peer
  * @memberof module:net/peer
  */
-export class Peer extends EventEmitter {
+export abstract class Peer extends EventEmitter {
   public config: Config
   public id: string
   public address: string
   public inbound: boolean
   public server: Server | undefined
-  public bound: Map<string, BoundProtocol>
   protected transport: string
   protected protocols: Protocol[]
   private _idle: boolean
@@ -58,9 +57,14 @@ export class Peer extends EventEmitter {
   public pooled: boolean = false
 
   // Dynamically bound protocol properties
+  // TODO: interfaces
   public eth: (BoundProtocol & EthProtocolMethods) | undefined
   public snap: (BoundProtocol & SnapProtocolMethods) | undefined
   public les: (BoundProtocol & LesProtocolMethods) | undefined
+
+  // Dynamic `peerScore` used to score peers by `best()`
+  // For `eth` protocol, this is the chain height of the peer
+  public peerScore: bigint
 
   /**
    * Create new peer
@@ -75,9 +79,10 @@ export class Peer extends EventEmitter {
     this.transport = options.transport
     this.inbound = options.inbound ?? false
     this.protocols = options.protocols ?? []
-    this.bound = new Map()
 
     this._idle = true
+
+    this.peerScore = BIGINT_NEG1
   }
 
   /**
@@ -94,59 +99,21 @@ export class Peer extends EventEmitter {
     this._idle = value
   }
 
-  async connect(): Promise<void> {}
-
-  /**
-   * Adds a protocol to this peer given a sender instance. Protocol methods
-   * will be accessible via a field with the same name as protocol. New methods
-   * will be added corresponding to each message defined by the protocol, in
-   * addition to send() and request() methods that takes a message name and message
-   * arguments. send() only sends a message without waiting for a response, whereas
-   * request() also sends the message but will return a promise that resolves with
-   * the response payload.
-   * @param protocol protocol instance
-   * @param sender sender instance provided by subclass
-   * @example
-   * ```typescript
-   * await peer.bindProtocol(ethProtocol, sender)
-   * // Example: Directly call message name as a method on the bound protocol
-   * const headers1 = await peer.eth.getBlockHeaders({ block: BigInt(1), max: 100 })
-   * // Example: Call request() method with message name as first parameter
-   * const headers2 = await peer.eth.request('getBlockHeaders', { block: BigInt(1), max: 100 })
-   * // Example: Call send() method with message name as first parameter and
-   * // wait for response message as an event
-   * peer.eth.send('getBlockHeaders', { block: BigInt(1), max: 100 })
-   * peer.eth.on('message', ({ data }) => console.log(`Received ${data.length} headers`))
-   * ```
-   */
-  protected async bindProtocol(protocol: Protocol, sender: Sender): Promise<void> {
-    const bound = await protocol.bind(this, sender)
-    this.bound.set(bound.name, bound)
-  }
-
-  /**
-   * Return true if peer understand the specified protocol name
-   * @param protocolName
-   */
-  understands(protocolName: string): boolean {
-    return !!this.bound.get(protocolName)
-  }
+  abstract connect(): Promise<void>
 
   /**
    * Handle unhandled messages along handshake
    */
-  handleMessageQueue() {
-    for (const bound of this.bound.values()) {
-      bound.handleMessageQueue()
-    }
-  }
+  abstract handleMessageQueue(): Promise<void>
+
+  abstract availableProtocols(): string[]
 
   toString(withFullId = false): string {
     const properties = {
       id: withFullId ? this.id : this.id.substr(0, 8),
       address: this.address,
       transport: this.transport,
-      protocols: Array.from(this.bound.keys()),
+      protocols: this.availableProtocols,
       inbound: this.inbound,
     }
     return Object.entries(properties)
