+++ packages/client/src/net/protocol/protocol.ts
@@ -1,4 +1,4 @@
-import { BoundProtocol } from './boundprotocol'
+import { Event } from '../../types'
 
 import type { Config } from '../../config'
 import type { Peer } from '../peer/peer'
@@ -9,6 +9,15 @@ export interface ProtocolOptions {
 
   /* Handshake timeout in ms (default: 8000) */
   timeout?: number
+
+  /* Protocol */
+  protocol: Protocol
+
+  /* Peer */
+  peer: Peer
+
+  /* Sender */
+  sender: Sender
 }
 
 export type Message = {
@@ -43,6 +52,15 @@ export class Protocol {
   public timeout: number
   public opened: boolean
 
+  public name: string
+  private protocol: Protocol
+  private peer: Peer
+  private sender: Sender
+  public versions: number[]
+  private _status: any
+  private resolvers: Map<string | number, any>
+  private messageQueue: Message[] = []
+
   /**
    * Create new protocol
    */
@@ -50,7 +68,159 @@ export class Protocol {
     this.config = options.config
     this.timeout = options.timeout ?? 8000
 
+    this.config = options.config
+
+    this.protocol = options.protocol
+    this.peer = options.peer
+    this.sender = options.sender
+    this.name = this.protocol.name
+    this.versions = this.protocol.versions
+    this.timeout = this.protocol.timeout
+    this._status = {}
+    this.resolvers = new Map()
+    this.sender.on('message', (message: any) => {
+      try {
+        if (this.peer.pooled) {
+          this.handle(message)
+        } else {
+          this.messageQueue.push(message)
+          // Expected message queue growth is in the single digits
+          // so this adds a guard here if something goes wrong
+          if (this.messageQueue.length >= 50) {
+            const error = new Error('unexpected message queue growth for peer')
+            this.config.events.emit(Event.PROTOCOL_ERROR, error, this.peer)
+          }
+        }
+      } catch (error: any) {
+        this.config.events.emit(Event.PROTOCOL_ERROR, error, this.peer)
+      }
+    })
+    this.sender.on('error', (error: Error) =>
+      this.config.events.emit(Event.PROTOCOL_ERROR, error, this.peer)
+    )
+
     this.opened = false
+  }
+
+  get status(): any {
+    return this._status
+  }
+
+  set status(status: any) {
+    Object.assign(this._status, status)
+  }
+
+  /**
+   * Handle incoming message
+   * @param message message object
+   * @emits {@link Event.PROTOCOL_MESSAGE}
+   * @emits {@link Event.PROTOCOL_ERROR}
+   */
+  private handle(incoming: Message) {
+    const messages = this.protocol.messages
+    const message = messages.find((m) => m.code === incoming.code)
+    if (!message) {
+      return
+    }
+
+    let data
+    let error
+    try {
+      data = this.protocol.decode(message, incoming.payload)
+    } catch (e: any) {
+      error = new Error(`Could not decode message ${message.name}: ${e}`)
+    }
+    const resolver = this.resolvers.get(incoming.code)
+    if (resolver !== undefined) {
+      clearTimeout(resolver.timeout)
+      this.resolvers.delete(incoming.code)
+      if (error) {
+        resolver.reject(error)
+      } else {
+        resolver.resolve(data)
+      }
+    } else {
+      if (error) {
+        this.config.events.emit(Event.PROTOCOL_ERROR, error, this.peer)
+      } else {
+        this.config.events.emit(
+          Event.PROTOCOL_MESSAGE,
+          { name: message.name, data },
+          this.protocol.name,
+          this.peer
+        )
+      }
+    }
+  }
+
+  /**
+   * Handle unhandled messages along handshake
+   */
+  handleMessageQueue() {
+    for (const message of this.messageQueue) {
+      this.handle(message)
+    }
+  }
+
+  /**
+   * Send message with name and the specified args
+   * @param name message name
+   * @param args message arguments
+   */
+  send(name: string, args?: any) {
+    const messages = this.protocol.messages
+    const message = messages.find((m) => m.name === name)
+    if (message) {
+      const encoded = this.protocol.encode(message, args)
+      this.sender.sendMessage(message.code, encoded)
+    } else {
+      throw new Error(`Unknown message: ${name}`)
+    }
+    return message
+  }
+
+  /**
+   * Returns a promise that resolves with the message payload when a response
+   * to the specified message is received
+   * @param name message to wait for
+   * @param args message arguments
+   */
+  async request(name: string, args: any[]): Promise<any> {
+    const message = this.send(name, args)
+    let lock
+    if (
+      typeof message.response === 'number' &&
+      this.resolvers.get(message.response) !== undefined
+    ) {
+      const res = this.resolvers.get(message.response)
+      lock = res.lock
+      await res.lock.acquire()
+    }
+    const resolver: any = {
+      timeout: null,
+      resolve: null,
+      reject: null,
+      lock: lock ?? new Lock(),
+    }
+    this.resolvers.set(message.response!, resolver)
+    if (lock === undefined) {
+      await resolver.lock.acquire()
+    }
+    return new Promise((resolve, reject) => {
+      resolver.resolve = function (e: any) {
+        resolver.lock.release()
+        resolve(e)
+      }
+      resolver.reject = function (e: any) {
+        resolver.lock.release()
+        reject(e)
+      }
+      resolver.timeout = setTimeout(() => {
+        resolver.timeout = null
+        this.resolvers.delete(message.response!)
+        resolver.reject(new Error(`Request timed out after ${this.timeout}ms`))
+      }, this.timeout)
+    })
   }
 
   /**
@@ -74,7 +244,7 @@ export class Protocol {
       const handleStatus = (status: any) => {
         if (timeout !== null && timeout !== 0) {
           clearTimeout(timeout)
-          resolve(this.decodeStatus(status))
+          this._status = this.decodeStatus(status)
         }
       }
       if (sender.status !== undefined && sender.status !== null && sender.status !== 0) {
@@ -83,20 +253,6 @@ export class Protocol {
         sender.once('status', handleStatus)
       }
     })
-  }
-
-  /**
-   * Abstract getter for name of protocol
-   */
-  get name() {
-    return 'protocol'
-  }
-
-  /**
-   * Protocol versions supported
-   */
-  get versions(): number[] {
-    throw new Error('Unimplemented')
   }
 
   /**
@@ -152,6 +308,7 @@ export class Protocol {
    * @param peer peer
    * @param sender sender
    */
+  /* TODO FIXME
   async bind(peer: Peer, sender: Sender): Promise<BoundProtocol> {
     const bound = new BoundProtocol({
       config: this.config,
@@ -168,5 +325,5 @@ export class Protocol {
     //@ts-ignore TODO: evaluate this line
     peer[this.name] = bound
     return bound
-  }
+  }*/
 }
